backend/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-ratelimit==4.1.0
drf-spectacular==0.27.2
Pillow==10.3.0
